Sign up expiration handling
