+++ backend/core/src/main/java/com/feed_grabber/core/invitation/dto/InvitationSignUpDto.java
@@ -11,5 +11,6 @@ public class InvitationSignUpDto {
     private String email;
     private String companyName;
     private Date createdAt;
+    private Boolean expired;
     private Boolean accepted;
 }
